# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MentorNepal API",
    "version": "1.0.0",
    "description": "API docs for MentorNepal backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "VerificationRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "mentorId": {
            "type": "string"
          },
          "mentorName": {
            "type": "string"
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewerNotes": {
            "type": "string"
          }
        }
      },
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "mentee",
              "mentor",
              "admin"
            ]
          },
          "onboardingComplete": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PlatformStats": {
        "type": "object",
        "properties": {
          "totalUsers": {
            "type": "integer"
          },
          "totalMentors": {
            "type": "integer"
          },
          "totalMentees": {
            "type": "integer"
          },
          "totalSessions": {
            "type": "integer"
          },
          "completedSessions": {
            "type": "integer"
          },
          "pendingVerifications": {
            "type": "integer"
          },
          "averageSessionRating": {
            "type": "number"
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/admin/verification-requests": {
      "get": {
        "summary": "Get pending verification requests",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ],
              "default": "pending"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of verification requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/VerificationRequest"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (admin only)"
          }
        }
      }
    },
    "/api/admin/verification-requests/{id}": {
      "patch": {
        "summary": "Approve or reject verification request",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "approved",
                      "rejected"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification request updated"
          },
          "400": {
            "description": "Invalid status"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (admin only)"
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "Get all users (admin)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "mentee",
                "mentor",
                "admin"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminUser"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (admin only)"
          }
        }
      }
    },
    "/api/admin/stats": {
      "get": {
        "summary": "Get platform statistics (cached)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Platform statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformStats"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (admin only)"
          }
        }
      }
    },
    "/api/admin/mentors/{mentorId}/suspend": {
      "patch": {
        "summary": "Suspend or unsuspend a mentor",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "mentorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "suspended"
                ],
                "properties": {
                  "suspended": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mentor status updated"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (admin only)"
          },
          "404": {
            "description": "Mentor not found"
          }
        }
      }
    },
...
  },
...
}